requirements.txt
PyQt5
Pillow
sqlite3
multiprocessing